[REFACTOR]: api위치 변경
